general improvements

+++ RoamerPipeline.py
@@ -24,21 +24,20 @@ MVG_PATH = "~/openMVG_build/Linux-x86 64-RELEASE/"
 MVS_PATH = "~/openMVS_build/bin/"
 
 #Get Directories
-INPUT_DIR = input("Enter the Input Directory")
-OUTPUT_DIR = input("Enter the Output Directory")
+INPUT_DIR = str(raw_input("Enter the Input Directory"))
+OUTPUT_DIR = str(raw_input("Enter the Output Directory"))
 
 while(True):
-    if(input("Are you happy with these directories? (Y/N)") == "Y"):
+    if(raw_input("Are you happy with these directories? (Y/N)") == "Y"):
         break
     time.sleep(1)
 
 if(OUTPUT_DIR[len(OUTPUT_DIR)-1] != "/"):
     OUTPUT_DIR = OUTPUT_DIR + "/"
 
-fileOut = open(OUTPUT_DIR + "/LOGS/pythonLog.txt", "w")
-
 #Create needed file structure
-os.rmdir(OUTPUT_DIR + "MVG")
+if(os.path.exists(OUTPUT_DIR + "MVG")):
+    os.rmdir(OUTPUT_DIR + "MVG")
 os.mkdir(OUTPUT_DIR + "MVG")
 prnt("Created MVG Dir")
 
@@ -46,11 +45,13 @@ os.mkdir(OUTPUT_DIR + "MVG/matches")
 os.mkdir(OUTPUT_DIR + "MVG/reconstruction_sequential")
 prnt("Populated MVG Dir")
 
-os.rmdir(OUTPUT_DIR + "MVS")
+if(os.path.exists(OUTPUT_DIR + "MVS")):
+    os.rmdir(OUTPUT_DIR + "MVS")
 os.mkdir(OUTPUT_DIR + "MVS")
 prnt("Created MVS Dir")
 
-os.rmdir(OUTPUT_DIR + "LOGS")
+if(os.path.exists(OUTPUT_DIR + "LOGS")):
+    os.rmdir(OUTPUT_DIR + "LOGS")
 os.mkdir(OUTPUT_DIR + "LOGS")
 prnt("Created LOGS Dir")
 
@@ -58,6 +59,7 @@ MVG_MATCHES = OUTPUT_DIR + "matches"
 MVG_RECONSTRUCT = OUTPUT_DIR + "reconstruction_sequential"
 MVS_OUTPUT = OUTPUT_DIR + "MVS"
 
+fileOut = open(OUTPUT_DIR + "LOGS/pythonLog.txt", "w")
 
 #START OF MVG PIPELINE
 prnt("1. Intrinsics Analysis")
